UsbAPI: add getDevices function
For transferring info about usb devices to libtermux-api in a format
that is easier to parse than json. Returns all the information
available through the android usbmanager that correspond to entries in
the libusb_device struct.

We can then create a list of libusb_device devices in libtermux-api
and return a pointer to it to programs, like a (poor) replacement of
libusb_get_device_list.